+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @c15t/nextjs
 
+## 1.8.3-canary-20251222100111
+
+### Patch Changes
+
+- @c15t/react@1.8.3-canary-20251222100111
+
+## 1.8.3-canary-20251218133143
+
+### Patch Changes
+
+- Updated dependencies [226f45c]
+  - @c15t/react@1.8.3-canary-20251218133143
+
 ## 1.8.2
 
 ### Patch Changes